Display ETH addresses using EIP55 mixed-case address encoding.
This is completely untested. I didn't see any existing examples of how to write
a testcase (not that I'm able to actually run anything yet at this point in my
"build it with clang" adventure). That being said, it does build for me.

https://github.com/ethereum/EIPs/blob/master/EIPS/eip-55.md

It'd be a good idea to try out the example testcases in the EIP before merging
this, but I'm not sure how to do that without being able to run the code. I
should get there eventually...