[emulator] Build a host-version of keepkey for native unit testing.
This adds a rough cut at a CMake build system, and cuts a few corners w.r.t.
getting a working "emulator". By that I mean there is still a bunch of
engineering work left to be done before we can start emulating the KK on the
host.  As an incremental step toward that, this patch gives us a way to write
unit tests to exercise functions in the firmware.  Additionally, it fully
replaces the SCons build system with CMake.

The first testcase checks that EIP55 checksum addresses are calculated
correctly for a few of the test vectors given in the EIP itself.

To build it, install CMake and do:

    $ mkdir build && cd build
    $ cmake -C path/to/keepkey-firmware/caches/emulator.cmake path/to/keepkey-firmware
    $ make

Build with sanitizers:

    $ mkdir build && cd build
    $ cmake -C path/to/keepkey-firmware/caches/emulator.cmake path/to/keepkey-firmware \
        -DCMAKE_C_FLAGS=-fsanitize=address \
        -DCMAKE_CXX_FLAGS=-fsanitize=address
    $ make

If your host compiler includes sanitizer support, you can build with:

  $ cmake path/to/keepkey-firmware \
    -DKK_EMULATOR=ON \
    -DCMAKE_BUILD_TYPE=Debug \
    -DCMAKE_CXX_FLAGS=-fsanitize=address \
    -DCMAKE_C_FLAGS=-fsanitize=address

And the compiler will insert some memory safety checks that are executed at
runtime.

To run the unittests, do:

    $ make test

To build the debug versions of the keepkey/salt firmwares, do:

    $ ./docker_build_firmware_debug.sh

likewise, to build the release versions:

    $ ./docker_build_firmware_release.sh