Revert "Constant time checks of the session pin"
This reverts commit f910dbf334da418b4cd0b894660e08366d4fd40d.

It was broken for cases where the storage pin contained something like:

    "1234\06789"

which happens when a user sets a long pin, then sometime later sets a pin that
is shorter by at least two digits.  In that case, the session pin checks
incorrectly diagnose a pin entry of "1234\0\0\0\0\0\0" as being "invalid",
since the digits after the null terminator were being compared, and used as
part of the correctness check.