introduce seedless setup (aka no_backup)
With the no_backup flag enabled, the recovery sentence will not be shown during
initialization. User assumes all risks associated with using this mode,
including but not limited to the fact that under very rare circumstances, the
seed words maybe be wiped during an unsuccessful firmware update.

This recovery sentence cannot be retrieved after the device has been
initialized, making the get_features message a reliable indicator of whether or
not the recovery sentence could have been shown during initialization.

This feature is for ADVANCED USERS ONLY. Please take the time to fully
understand the risks before using it. It is highly recommended that this policy
only be used with some other backup strategy involving e.g. 2-of-3 multisig
where the other two keys are properly backed up in cold storage.

This is a re-hash of firmware-private #41 "Implement a 'NoRecovery' policy."
from back in March, which got left on the back-burner and never got merged.

Inspired by:
https://www.reddit.com/r/keepkey/comments/7zfumd/can_i_configure_keepkey_to_initialize_without/
https://github.com/trezor/trezor-mcu/commit/4a0f0259d131f3f2681ad7540f4fe797e89429ff