bootloader: don't read secrets into ram during firmware update.
Due to a disclosure at 35c3 by Dmitry Nedospasov, Josh Datko and Thomas Roth,
we cannot load read the secrets into ram anymore during firmware update. To
mitigate this, and still allow preservation of secrets across firmware updates,
we check for the presence of some magic bytes at the beginning of the storage
sector immediately following the active sector.  If the bytes do not match the
known magic, secrets are wiped before booting.