u2f: fix #102 'U2F webauthn acts flakey in chrome'
One thing of note in this patch: button animation behaves slightly differently
than the rest of the confirm dialogs, as the animation happens *afer* the
action instead of just before. This is necessary because the U2F spec demands
answers to requests before our animation would have finished. As a "good
enough" measure, instead we send the response at the moment when the user
starts pressing the button, instead of when they release at the end of the
animation.

Fixes #102